# api-specification.yaml
openapi: 3.0.1
info:
  title: HCX Mock & Provider Service API's
  description: >-
    HCX Mockservice
  version: "1.0.0"
servers:
  - url: "https://api.example.com"
    description: Mock server
...
paths:
...
  /user/create:
    post:
      tags:
      - User Management API's
      summary: Create a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile:
                  type: string
                  description: The user's mobile number.
                name:
                  type: string
                  description: The user's name.
                address:
                  type: string
                  description: The user's address.
                payor_details:
                  type: object
                  description: Additional details about the user's payor.
                medical_history:
                  type: object
                  description: Information about the user's medical history.
              required:
                - mobile
      responses:
        '200':
          description: Successful operation. Returns user information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  timestamp:
                    type: integer
                    format: int64
                    description: The timestamp of the response.
                  mobile:
                    type: string
                    description: The user's mobile number.
...
  /user/search:
    get:
      tags:
        - User Management API's
      summary: Search for a user by mobile number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile:
                  type: string
                  description: The user's mobile number.
      responses:
        '200':
          description: Successful operation. Returns user information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        userName:
          type: string
          description: The name of the user.
        beneficiaryId:
          type: string
          description: The beneficiary ID of the user.
        address:
          type: string
          description: The address of the user.
        payorDetails:
          type: object
          description: Additional details about the user's payor.
        medicalHistory:
          type: object
          description: Information about the user's medical history.  
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT